+++ tests/routerCommunicator.cpp
@@ -5,6 +5,7 @@
 #include <cstddef>
 #include <cmath>
 #include <thread>
+#include <array>
 
 #include "util.h"
 
@@ -55,7 +56,7 @@ TEST_CASE("TransparentOutputStreamCommunicator - recipients", "[routerCommunicat
     std::vector<std::unique_ptr<BufferTransmitter>> transmitters;
     std::vector<std::unique_ptr<Router::Handle>> handles;
 
-    for (size_t i = 0; i < packetSizes.size(); i++) {
+    for (uint8_t i = 0; i < packetSizes.size(); i++) {
         transmitters.push_back(std::make_unique<BufferTransmitter>(packetSizes[i]));
         handles.push_back(std::make_unique<Router::Handle>(router.subscribeTx(i, *transmitters.back())));
     }
@@ -402,7 +403,7 @@ TEST_CASE("Multithread UnboundedBufferedInputPacketCommunicator", "[routerCommun
     router.subscribeChannel(CHANNEL, communicator);
 
     std::vector<std::vector<uint8_t>> packets;
-    for (size_t i = 0; i < 100; i++) {
+    for (uint8_t i = 0; i < 100; i++) {
         packets.push_back(rangeVector<uint8_t>(i, PACKET_SIZE));
     }
